requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2022.9.24
chardet==5.1.0
charset-normalizer==3.0.1
colorzero==2.0
distro==1.8.0
dotenv==0.9.9
elevenlabs==2.1.0
gpiozero==2.0.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.3
jiter==0.10.0
lgpio==0.2.2.0
openai==1.82.0
pigpio==1.78
pycryptodomex==3.11.0
pydantic==2.11.5
pydantic_core==2.33.2
python-apt==2.6.0
python-dotenv==1.1.0
requests==2.28.1
rpi-lgpio==0.6
six==1.16.0
smbus2==0.4.2
sniffio==1.3.1
spidev==3.5
ssh-import-id==5.10
toml==0.10.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==1.26.12
websockets==15.0.1
